backend/requirements.txt
flask
flask-cors
python-dotenv
pandas
numpy<2.0.0
requests
python-dateutil
supabase>=0.7.1
psycopg2-binary
python-bcb==0.3.3
yfinance
matplotlib
seaborn
